+++ lib/spior/service/enable.rb
@@ -19,6 +19,8 @@ module Spior
           for_arch
         when :void
           for_void
+        when :debian
+          for_debian
         else
           Msg.report 'Your distro is not yet supported.'
         end
@@ -27,19 +29,19 @@ module Spior
       protected
 
       def for_gentoo
-        Iptables::Rules.new.save
         case Nomansland.init?
         when :systemd
+          Iptables::Rules.new.save
           systemd_enable('iptables-restore', 'tor')
         when :openrc
-          rc_upd = Helpers::Exec.new('rc-update')
-          rc_upd.run('rc-update add iptables boot')
-          rc_upd.run('rc-update add tor')
-          rc_upd.run('rc-update add tor default')
+          Helpers.cmd('rc-service iptables save')
+          Helpers.cmd('rc-update add iptables boot') # default or boot ?
+          Helpers.cmd('rc-update add tor default')
         else
           Msg.report 'Init no yet supported for start Iptables at boot'
         end
         Ipv6.new.block_persist
+        Msg.p 'Persist enabled for Gentoo...'
       end
 
       def for_arch
@@ -58,14 +60,26 @@ module Spior
         Msg.p 'Persist enabled for Void...'
       end
 
+      # https://wiki.debian.org/iptables
+      def for_debian
+        Iptables::Rules.new.save
+        Tor::Config.new(Tempfile.new('torrc')).backup
+        systemd_enable('tor')
+        File.write('/tmp/start_iptables',
+                   "#!/bin/sh\n/sbin/iptables-restore < /etc/iptables.up.rules\n")
+        Helpers.mv('/tmp/start_iptables', '/etc/network/if-pre-up.d/iptables')
+        Helpers.cmd('chmod +x /etc/network/if-pre-up.d/iptables')
+        Ipv6.new.block_persist
+        Msg.p 'Persist mode enabled for Debian...'
+      end
+
       private
 
       def systemd_enable(*services)
-        systemctl = Helpers::Exec.new('systemctl')
         services.each do |s|
           next if system("systemctl is-enabled #{s} >/dev/null")
 
-          systemctl.run("enable #{s}")
+          Helpers.cmd("systemctl enable #{s}")
           Msg.p "Enabling #{s}..."
         end
       end
@@ -74,17 +88,16 @@ module Spior
         services.each do |s|
           next if File.exist? "/var/service/#{s}"
 
-          Helpers::Exec.new('ln').run("-s /etc/sv/#{s} /var/service/#{s}")
+          Helpers.cmd("ln -s /etc/sv/#{s} /var/service/#{s}")
           Msg.p "Enabling #{s}"
         end
       end
 
       def systemd_start(service)
-        systemctl = Helpers::Exec.new('systemctl')
         return if system("systemctl is-active #{service} >/dev/null")
 
         Msg.p "Search for service #{service}..."
-        systemctl.run("start #{service}")
+        Helpers.cmd("systemctl start #{service}")
       end
     end
   end